+++ src/components/button/mod.rs
@@ -1,3 +1,3 @@
 pub mod component;
 
-pub use component::{BASIC_BUTTONS, Button, ButtonKind, ICON_BUTTONS};
+pub use component::{BeverlyButton, ButtonColor, ButtonPlugin};